Aesthetics revamp
The entire page has been revamped. Now the bookmarks (including those in
collapsible groups) look like buttons, the favicons are automatically
retrieved from the links and shown inside the bookmark, and the hovering
effect is different.

In addition to that, the colors have been modified and now the page has
a dark mode.